requirements.txt
google-generativeai
streamlit
pandas
openpyxl
python-docx
pypdf
pillow
python-dotenv